chore: unify mock data duplicates
