+++ packages/testland/e2e/testcases/application/workqueue-flow-2.spec.ts
@@ -14,67 +14,24 @@ import {
   continueForm,
   drawSignature,
   formatName,
-  getRandomDate,
   goToSection,
   login,
   triggerDeclarationAction
 } from '../../helpers'
 import { CREDENTIALS } from '../../constants'
 import { ensureAssignedToUser, selectAction } from '../../utils'
-import { assertRecordInWorkqueue, fillDate } from '../birth/helpers'
+import {
+  assertRecordInWorkqueue,
+  fillDate,
+  generateBirthInputs
+} from '../birth/helpers'
 import { openRecordByTitle } from '../print-certificate/birth/helpers'
 
 // HO Notifies => Registrar Registers
 
 test.describe.serial('2. Workqueue flow - 2', () => {
   let page: Page
-  const declaration = {
-    child: {
-      name: {
-        firstNames: faker.person.firstName('male'),
-        familyName: faker.person.lastName('male')
-      },
-      gender: 'Male',
-      birthDate: getRandomDate(0, 200)
-    },
-    placeOfBirth: 'Health Institution',
-    birthLocation: { facility: 'Klow Village Hospital' },
-    informantType: 'Mother',
-    informantEmail: faker.internet.email(),
-    mother: {
-      name: {
-        firstNames: faker.person.firstName('female'),
-        familyName: faker.person.lastName('female')
-      },
-      birthDate: getRandomDate(20, 200),
-      nationality: 'Farajaland',
-      identifier: {
-        id: faker.string.numeric(10),
-        type: 'National ID'
-      },
-      address: {
-        country: 'Farajaland',
-        province: 'Sulaka',
-        district: 'Irundu',
-        village: 'Xhosa'
-      }
-    },
-    father: {
-      name: {
-        firstNames: faker.person.firstName('male'),
-        familyName: faker.person.lastName('male')
-      },
-      birthDate: getRandomDate(22, 200),
-      nationality: 'Gabon',
-      identifier: {
-        id: faker.string.numeric(10),
-        type: 'National ID'
-      },
-      address: {
-        sameAsMother: true
-      }
-    }
-  }
+  const declaration = generateBirthInputs()
   test.beforeAll(async ({ browser }) => {
     page = await browser.newPage()
   })
