+++ packages/testland/e2e/testcases/application/workqueue-flow-5.spec.ts
@@ -14,68 +14,24 @@ import {
   continueForm,
   drawSignature,
   formatName,
-  getRandomDate,
   goToSection,
   login,
   triggerDeclarationAction
 } from '../../helpers'
 import { CREDENTIALS } from '../../constants'
 import { ensureAssignedToUser, selectAction } from '../../utils'
-import { assertRecordInWorkqueue, fillDate } from '../birth/helpers'
+import {
+  assertRecordInWorkqueue,
+  fillDate,
+  generateBirthInputs
+} from '../birth/helpers'
 import { openRecordByTitle } from '../print-certificate/birth/helpers'
 
 // HO Notifies => RO Rejects => RO Declares and validates => Registrar rejects
 // => RO Re-declares again => Registrar registers
 test.describe.serial('5. Workqueue flow - 5', () => {
   let page: Page
-  const declaration = {
-    child: {
-      name: {
-        firstNames: faker.person.firstName('male'),
-        familyName: faker.person.lastName('male')
-      },
-      gender: 'Male',
-      birthDate: getRandomDate(0, 200)
-    },
-    placeOfBirth: 'Health Institution',
-    birthLocation: { facility: 'Klow Village Hospital' },
-    informantType: 'Mother',
-    informantEmail: faker.internet.email(),
-    mother: {
-      name: {
-        firstNames: faker.person.firstName('female'),
-        familyName: faker.person.lastName('female')
-      },
-      birthDate: getRandomDate(20, 200),
-      nationality: 'Farajaland',
-      identifier: {
-        id: faker.string.numeric(10),
-        type: 'National ID'
-      },
-      address: {
-        country: 'Farajaland',
-        province: 'Sulaka',
-        district: 'Irundu',
-        village: 'Xhosa'
-      }
-    },
-    father: {
-      name: {
-        firstNames: faker.person.firstName('male'),
-        familyName: faker.person.lastName('male')
-      },
-      birthDate: getRandomDate(22, 200),
-      nationality: 'Gabon',
-      identifier: {
-        id: faker.string.numeric(10),
-        type: 'National ID'
-      },
-      address: {
-        sameAsMother: true
-      }
-    }
-  }
-
+  const declaration = generateBirthInputs()
   const childName = formatName(declaration.child.name)
 
   test.beforeAll(async ({ browser }) => {
